pyproject.toml
[project]
name = "dify-agent-tool"
version = "0"
requires-python = ">=3.12"
dependencies = [
    "dify-plugin",
    "davidkhala.utils[http_request]",
    "davidkhala.ai",
]

[dependency-groups]
dev = [
    "pytest"
]

[tool.hatch.build.targets.wheel]
packages = ["tools"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


